requirement.txt
streamlit
langchain==0.1.0
langchain-community
python-dotenv